Add textHelpers to export text helper functions
