+++ src/helpers/textHelpers.js
@@ -1,0 +1,33 @@
+import React from 'react';
+
+export const capitalize = str => str.charAt(0).toUpperCase() + str.slice(1);
+
+export const generateParagraphs = text => {
+  const lines = text.split('. ');
+  let temp = '';
+  const paragraphs = [];
+
+  for (let line of lines) {
+    if (temp.length + line.length > 300) {
+      paragraphs.push(temp);
+      temp = line + '. ';
+    } else {
+      temp += line + '. ';
+    }
+  }
+  paragraphs.push(temp.substring(0, temp.length - 2));
+
+  return (
+    <div
+      style={{
+        margin: '0 auto',
+        maxWidth: '600px',
+        fontSize: '1.2em',
+      }}
+    >
+      {paragraphs.map((line, i) => (
+        <p key={i}>{line}</p>
+      ))}
+    </div>
+  );
+};
